packages/rfm_policy/pyproject.toml
[project]
name = "robot-rfm-policy"
version = "0.1.0"
description = "Robot Foundation Model (RFM) based policy learning package"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "torch>=1.11.0",
    "torchvision>=0.12.0",
    "numpy>=1.24.4",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "wandb>=0.12.0",
    "tqdm>=4.64.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.10.1",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.12.3",
    "jupyter-client>=8.6.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
